requirements.txt
addict==2.2.1
albumentations==0.4.0
autopep8==1.5.2
beautifulsoup4==4.8.1
blis==0.4.1
Bottleneck==1.2.1
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
cloudpickle==1.2.2
colorama==0.4.3
crccheck==0.6
cycler==0.10.0
cymem==2.0.2
cytoolz==0.10.0
dask==2.4.0
dataclasses==0.7
decorator==4.4.0
deep-pipe==0.0.7.post1
docopt==0.6.2
dominate==2.4.0
easydict==1.9
fastai==1.0.59
fastprogress==0.1.21
geomloss==0.2.3
idna==2.8
imageio==2.5.0
imgaug==0.2.6
importlib==1.0.4
importlib-metadata==0.23
ipython3==0.0.1
joblib==0.13.2
jsonpatch==1.24
jsonpointer==2.0
kaggle==1.5.6
kiwisolver==1.1.0
lmdb==0.98
matplotlib==3.1.1
MedPy==0.4.0
mkl-fft==1.0.14
mkl-random==1.0.2
mkl-service==2.3.0
more-itertools==7.2.0
munch==2.3.2
murmurhash==1.0.2
networkx==2.3
nibabel==2.5.1
numexpr==2.7.0
numpy==1.16.5
nvidia-ml-py3==7.352.0
olefile==0.46
opencv-python==4.1.1.26
packaging==19.2
pandas==0.25.1
pdp==0.3.0
Pillow==6.1.0
pipreqs==0.4.10
plac==1.1.3
preshed==3.0.2
pretrainedmodels==0.7.4
pretty-errors==1.2.7
protobuf==3.11.2
pycodestyle==2.5.0
pycparser==2.19
pydicom==1.3.0
pynvml==8.0.4
pyparsing==2.4.2
python-dateutil==2.8.0
python-slugify==3.0.6
pytz==2019.2
PyWavelets==1.0.3
PyYAML==5.1.2
pyzmq==18.1.1
requests==2.22.0
resource-manager==0.10.1
scikit-image==0.15.0
scikit-learn==0.21.2
scipy==1.3.1
setproctitle==1.1.10
SimpleITK==1.2.2
six==1.12.0
soupsieve==1.9.5
spacy==2.2.2
srsly==0.2.0
tensorboard-easy==0.2.3
text-unidecode==1.3
thinc==7.3.1
thop==0.0.31.post1910280903
toolz==0.10.0
torch==1.1.0
torchfile==0.1.0
torchvision==0.4.0a0+9232c4a
tornado==6.0.3
tqdm==4.32.1
urllib3==1.24.3
visdom==0.1.8.9
wasabi==0.4.0
websocket-client==0.57.0
yarg==0.1.9
zipp==0.6.0
